fix: StrongBoxEntity: Change decimal to double the Saldo and Ammount data

+++ Poliedro.Eds.Domain/StrongBox/Services/IStrongBoxService.cs
@@ -12,7 +12,7 @@ public interface IStrongBoxService
     Task<StrongBoxEntity> CreateAsync(
         long? idCorte,
         string type,
-        decimal ammount,
+        double ammount,
         string? note,
         CancellationToken cancellationToken);
 }
